Refactor : control 모듈, 시도군구 리스트 등 일부 기능 수정
API 통신 과정에 쓰이는 모듈의 위치와 시도군구 리스트에 대한 형식을 수정하였습니다.
