+++ src/components/main/Home/HomeBtnList.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom"
 
 import { AuthStateContext } from "module/Auth"
 import { leftToRight } from "styles/Animation"
-import { accountControl } from "api/accountControl"
+import { accountControl } from "api/controls/accountControl"
 
 const HomeBtnList = () => {
   const authState = useContext(AuthStateContext)
@@ -22,7 +22,6 @@ const HomeBtnList = () => {
       </HomeBtn>
       {authenticated ? (
         <>
-
           <HomeBtn to="/mypage/profile" delay={3}>
             마이페이지
           </HomeBtn>
