+++ src/components/main/Mypage/Profile/Modal/MypageBoxModal.jsx
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faXmark } from "@fortawesome/free-solid-svg-icons"
 import { modalShow } from "styles/Animation"
 import { ModalContext } from "module/Modal"
-import { mypageControl } from "api/mypageControl"
+import { mypageControl } from "api/controls/mypageControl"
 import defaultProfile from "assets/img/defaultProfile.png"
 const MypageBoxModal = ({ setPhoto }) => {
   const { closeModal } = useContext(ModalContext)
