--- src/constants/MainPageBottomText.js
@@ -1,7 +1,0 @@
-export const MainPageBottomText = [
-  { title: "Soil Bank", desc: "옥상녹화는 투수면적을 늘림으로 도시 홍수를 막을 수 있습니다." },
-  { title: "2", desc: "이것은 테스트 메세지입니다." },
-  { title: "3", desc: "이것은 테스트 메세지입니다." },
-  { title: "4", desc: "이것은 테스트 메세지입니다." },
-  { title: "5", desc: "이것은 테스트 메세지입니다." },
-]
