+++ src/constants/SidoGunguList.js
@@ -73,25 +73,15 @@ export const SidoGunguList = new Map([
   [
     "경기도",
     [
-      "수원시 장안구",
-      "수원시 권선구",
-      "수원시 팔달구",
-      "수원시 영통구",
-      "성남시 수정구",
-      "성남시 중원구",
-      "성남시 분당구",
+      "수원시",
+      "성남시",
       "의정부시",
-      "안양시 만안구",
-      "안양시 동안구",
+      "안양시",
       "부천시",
       "광명시",
       "평택시",
       "동두천시",
-      "안산시 상록구",
-      "안산시 단원구",
-      "고양시 덕양구",
-      "고양시 일산동구",
-      "고양시 일산서구",
+      "안산시",
       "과천시",
       "구리시",
       "남양주시",
@@ -100,9 +90,7 @@ export const SidoGunguList = new Map([
       "군포시",
       "의왕시",
       "하남시",
-      "용인시 처인구",
-      "용인시 기흥구",
-      "용인시 수지구",
+      "용인시",
       "파주시",
       "이천시",
       "안성시",
@@ -143,10 +131,7 @@ export const SidoGunguList = new Map([
   [
     "충청북도",
     [
-      "청주시 상당구",
-      "청주시 서원구",
-      "청주시 흥덕구",
-      "청주시 청원구",
+      "청주시",
       "충주시",
       "제천시",
       "보은군",
@@ -162,8 +147,7 @@ export const SidoGunguList = new Map([
   [
     "충청남도",
     [
-      "천안시 동남구",
-      "천안시 서북구",
+      "천안시",
       "공주시",
       "보령시",
       "아산시",
@@ -183,8 +167,7 @@ export const SidoGunguList = new Map([
   [
     "전라북도",
     [
-      "전주시 완산구",
-      "전주시 덕진구",
+      "전주시",
       "군산시",
       "익산시",
       "정읍시",
@@ -230,8 +213,7 @@ export const SidoGunguList = new Map([
   [
     "경상북도",
     [
-      "포항시 남구",
-      "포항시 북구",
+      "포항시",
       "경주시",
       "김천시",
       "안동시",
@@ -259,11 +241,7 @@ export const SidoGunguList = new Map([
   [
     "경상남도",
     [
-      "창원시 의창구",
-      "창원시 성산구",
-      "창원시 마산합포구",
-      "창원시 마산회원구",
-      "창원시 진해구",
+      "창원시",
       "진주시",
       "통영시",
       "사천시",
